+++ main/Size.pde
@@ -1,7 +1,7 @@
 public class Size{
-  public int breite;
-  public int hoehe;
-  public Size(int breite, int hoehe){
+  public float breite;
+  public float hoehe;
+  public Size(float breite, float hoehe){
     this.breite = breite;
     this.hoehe = hoehe;
   }
